+++ example/a.ts
@@ -1,77 +1,67 @@
-import { Elysia, t } from '../src'
-import { req } from '../test/utils'
+import { Elysia, ElysiaStatus, t, type UnwrapSchema } from '../src'
 
-const challengeModel = t.Object({
-	nonce: t.String(),
-	issued: t.Number(),
-	bits: t.Number()
-})
+const Models = {
+	'user.update': t.Object({
+		id: t.String(),
+		name: t.Optional(t.String())
+	})
+}
 
-const app = new Elysia({
-	cookie: {
-		secrets: ['a', null],
-		sign: 'challenge'
-	}
-})
-	.get(
-		'/set',
-		({ cookie: { challenge } }) => {
-			challenge.value = {
-				nonce: 'hello',
-				bits: 19,
-				issued: Date.now()
-			}
+type Models = {
+	[k in keyof typeof Models]: UnwrapSchema<(typeof Models)[k]>
+}
 
-			return challenge.value
-		},
-		{
-			cookie: t.Cookie({
-				challenge: t.Optional(challengeModel)
-			})
-		}
-	)
-	.get(
-		'/get',
-		({ cookie: { challenge } }) => {
+const app = new Elysia()
+	.macro('isAuth', {
+		// headers: t.Object({
+		// 	authorization: t.TemplateLiteral('Authorization ${string}')
+		// }),
+		async resolve({ headers, status }) {
+			// Mock authentication logic
+			if (Math.random() > 0.5) return status(401, 'Not signed in')
+			if (Math.random() > 0.5) return status(401, 'Deactivated account')
+
 			return {
-				type: typeof challenge,
-				value: challenge.value
+				user: 'saltyaom'
 			}
-		},
-		{
-			cookie: t.Cookie({
-				challenge: challengeModel
-			})
-		}
-	)
-
-const first = await app.handle(
-	req('/set', {
-		headers: {
-			cookie: `challenge=${JSON.stringify({
-				nonce: 'hello',
-				bits: 19,
-				issued: 1770750432990
-			})}`
 		}
 	})
-)
+	.model(Models)
+	.macro('isAdmin', {
+		isAuth: true,
+		async resolve({ headers, status }) {
+			// Mock admin check logic
+			if (Math.random() > 0.5) return status(403, 'Not allowed')
 
-console.log(first.status)
-console.log(await first.json())
 
-const cookie = first.headers.get('set-cookie')
-const challenge = cookie!.match(/challenge=([^;]*)/)![1]
 
-console.log(challenge)
+			headers.authorization
 
-const second = await app
-	.handle(
-		req('/get', {
-			headers: {
-				cookie: `challenge=${challenge}`
+			return {
+				admin: {
+					async updateUser({ id, ...rest }: Models['user.update']) {
+						if (Math.random() > 0.5) return status(404, 'No user')
+
+						return { id }
+					}
+				}
 			}
-		})
+		}
+	})
+	.post(
+		'/',
+		async ({ user, admin, body, headers }) => {
+			return 'ok'
+			// const updated = await admin.updateUser(body)
+
+			// if (updated instanceof ElysiaStatus) return updated
+
+			// return `User ${user} updated user ${updated.id}` as const
+		},
+		{
+			isAuth: true,
+			// body: 'user.update'
+		}
 	)
 
-console.log(second.status)
+type Result = (typeof app)['~Routes']['post']['response']